refactor: updated the filtering logic
